requirements.txt
flask==2.0.1
firebase-admin==6.2.0
google-generativeai==0.3.0
requests==2.26.0
python-dotenv==0.19.0
geopy==2.2.0

#pip install werkzeug==2.3.0
